Part3_Deployment_Demo/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.0
Pillow>=8.0.0
matplotlib>=3.3.0
opencv-python>=4.5.0
PyQt5>=5.15.0
